Update examples to match mip_interface changes and move some common code to example_utils.
